Add files via upload
I wish to loop through an entire array then filter the array based on the the shelves each book is arranged before mapping the result to display each book on its shelf. I created an IIFE to handle the logic of checking which book to display on which shelf but it isn't working. The code is found on the MainPage.js.

I am new to react and open to learn . All suggestion are appreciated.